requirements.txt
# Copyright 2023 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
# ==============================================================================
absl-py==2.0.0
nltk==3.8.1
numpy==1.25.2
google-generativeai==0.2.0
scipy==1.11.3
etils==1.5.0
simple-parsing==0.1.4
pandas==2.1.1
ratelimit==2.2.1
Flask==2.3.3